openclaw-nexus/nexusctl/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "nexusctl"
version = "0.1.0"
description = "OpenClaw Nexus control-plane implementation"
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0"]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["../tests", "tests"]

[project.scripts]
nexusctl = "nexusctl.interfaces.cli.main:main"
